+++ frontend-teesa/index.html
@@ -5,6 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Teesa</title>
+
   </head>
   <body>
     <div id="root"></div>
