+++ frontend-teesa/src/App.jsx
@@ -3,10 +3,12 @@ import { Route, Routes, useLocation } from 'react-router-dom';
 //Components
 import Home from './Components/Home/Home';
 import Navbar from './Components/Navbar/Navbar';
-import Landing from './Components/Landing/Landing';
+import Landing from './views/Landing/Landing';
 import Detail from './Components/Detail/Detail';
 import Login from './Components/Login/Login';
-import Services from './Components/Services/Services';
+import Services from './views/Services/Services';
+import Contact from './views/Contact/Contact';
+import Direccion from './views/Direccion/Direccion';
 
 function App() {
   const location = useLocation();
@@ -21,6 +23,9 @@ function App() {
         <Route path='/home/:id' element={<Detail />} />
         <Route path='/login' element={<Login />} />
         <Route path='/services' element={<Services />}/>
+        <Route path='/contact' element={<Contact />}/>
+        <Route path='/location' element={<Direccion/>}/>
+        
       </Routes>
     </div>
   );
