--- frontend-teesa/src/Components/Services/Services.jsx
@@ -1,8 +1,0 @@
-function Services() {
-    return (
-      <div></div>
-
-    );
-  }
-  
-  export default Services;
