+++ frontend-teesa/src/views/Direccion/Direccion.jsx
@@ -1,0 +1,8 @@
+function Direccion() {
+    return (
+      <div></div>
+
+    );
+  }
+  
+  export default Direccion;
