+++ frontend-teesa/src/views/Services/Services.jsx
@@ -1,0 +1,8 @@
+function Services() {
+    return (
+      <div></div>
+
+    );
+  }
+  
+  export default Services;
